@@ -1,13 +1,13 @@
 <template>
   <div>
-    <nuxt/>
+    <nuxt />
   </div>
 </template>
 
 <style>
-html
-{
-  font-family: "Source Sans Pro", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+html {
+  font-family: 'Source Sans Pro', -apple-system, BlinkMacSystemFont, 'Segoe UI',
+    Roboto, 'Helvetica Neue', Arial, sans-serif;
   font-size: 16px;
   word-spacing: 1px;
   -ms-text-size-adjust: 100%;
@@ -16,13 +16,15 @@ html
   -webkit-font-smoothing: antialiased;
   box-sizing: border-box;
 }
-*, *:before, *:after
-{
+
+*,
+*:before,
+*:after {
   box-sizing: border-box;
   margin: 0;
 }
-.button--green
-{
+
+.button--green {
   display: inline-block;
   border-radius: 4px;
   border: 1px solid #3b8070;
@@ -30,13 +32,13 @@ html
   text-decoration: none;
   padding: 10px 30px;
 }
-.button--green:hover
-{
+
+.button--green:hover {
   color: #fff;
   background-color: #3b8070;
 }
-.button--grey
-{
+
+.button--grey {
   display: inline-block;
   border-radius: 4px;
   border: 1px solid #35495e;
@@ -45,8 +47,8 @@ html
   padding: 10px 30px;
   margin-left: 15px;
 }
-.button--grey:hover
-{
+
+.button--grey:hover {
   color: #fff;
   background-color: #35495e;
 }
